+++ modules/foundry_script/tests/scripts/analyzer/features/typed_rest_parameter_external_provider.notest.fs
@@ -1,5 +1,5 @@
-func collect(prefix: String, ...values: Array[int]) -> int:
+func collect(prefix: String, ...values: Array[int]) -> long:
 	return prefix.length() + values.size()
 
-func callback() -> Callable[[String, ...Array[int]], int]:
+func callback() -> Callable[[String, ...Array[int]], long]:
 	return collect
